click handler update
